benchmark: consolidate FP8 GDN performance test
